moved interfaces from API\Abstractions into API\Interfaces to centralize. Interfaces appear to be a common folder in WPF/MVVM apps.
