+++ src/Asn1Editor/API/Interfaces/IHasAsnDocumentTabs.cs
@@ -4,7 +4,7 @@ using SysadminsLV.Asn1Editor.API.ModelObjects;
 using SysadminsLV.Asn1Editor.API.ViewModel;
 using SysadminsLV.Asn1Editor.Core.Tree;
 
-namespace SysadminsLV.Asn1Editor.API.Abstractions;
+namespace SysadminsLV.Asn1Editor.API.Interfaces;
 
 public interface IHasAsnDocumentTabs {
     UserSettings UserSettings { get; }
