+++ src/Asn1Editor/API/Interfaces/IOidEditorVM.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows.Input;
 
-namespace SysadminsLV.Asn1Editor.API.Abstractions;
+namespace SysadminsLV.Asn1Editor.API.Interfaces;
 
 public interface IOidEditorVM {
     ICommand ReloadCommand { get; }
