+++ src/Asn1Editor/API/ViewModel/NewAsnNodeEditorVM.cs
@@ -2,7 +2,7 @@
 using System.Globalization;
 using System.Text.RegularExpressions;
 using System.Windows.Input;
-using SysadminsLV.Asn1Editor.API.Abstractions;
+using SysadminsLV.Asn1Editor.API.Interfaces;
 using SysadminsLV.Asn1Editor.API.ModelObjects;
 using SysadminsLV.Asn1Parser;
 using SysadminsLV.Asn1Parser.Universal;
