+++ src/Asn1Editor/API/ViewModel/TextViewerVM.cs
@@ -2,7 +2,6 @@
 using System.Globalization;
 using System.IO;
 using System.Windows.Input;
-using SysadminsLV.Asn1Editor.API.Abstractions;
 using SysadminsLV.Asn1Editor.API.Interfaces;
 using SysadminsLV.Asn1Editor.API.ModelObjects;
 using SysadminsLV.Asn1Editor.Core.AsnFormatters;
